+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+### 2024-02-08
+
+### Changed
+
+- Updated the Item Lookup page filters to be used in the URL, allowing you to share a link to a specific set of tags with the filters applied.
+
 ## 2024-02-07
 
 ### Changed